+++ app/src/main/java/ru/practicum/android/diploma/features/favourite/data/repository/FavouriteVacanciesRepositoryImpl.kt
@@ -1,25 +1,29 @@
 package ru.practicum.android.diploma.features.favourite.data.repository
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flow
 import ru.practicum.android.diploma.features.common.data.database.AppDatabase
-import ru.practicum.android.diploma.features.common.data.database.VacancyDbEntity
-import ru.practicum.android.diploma.features.favourite.data.dto.toDomain
+import ru.practicum.android.diploma.features.common.data.database.KeySkillEntity
+import ru.practicum.android.diploma.features.favourite.data.dto.toDb
 import ru.practicum.android.diploma.features.favourite.domain.api.FavouriteVacanciesRepository
 import ru.practicum.android.diploma.features.favourite.domain.model.FavouriteVacancy
-import ru.practicum.android.diploma.features.search.domain.model.Vacancy
 
 class FavouriteVacanciesRepositoryImpl(
     private val appDatabase: AppDatabase,
 ) : FavouriteVacanciesRepository {
 
-    override fun getFavourites(): Flow<List<Vacancy>> = flow {
-        val vacancies = appDatabase.favouritesDao().getFavourites()
-        emit(convertFromDbEntity(vacancies))
+    override suspend fun addToFavourites(vacancy: FavouriteVacancy, keySkills: String) {
+        appDatabase.favouritesDao().addToFavourites(
+            vacancy.toDb(),
+            createKeySkillEntity(vacancy, keySkills)
+        )
     }
 
-    private fun convertFromDbEntity(vacanciesDbEntity: List<VacancyDbEntity>): List<FavouriteVacancy> {
-        return vacanciesDbEntity.map { it.toDomain()}
+    private fun createKeySkillEntity(vacancy: FavouriteVacancy, keySkills: String): KeySkillEntity {
+        return KeySkillEntity(
+            id = 0,
+            vacancyId = vacancy.vacancyId,
+            keySkill = keySkills,
+        )
     }
 
 }
+
