+++ app/src/main/java/ru/practicum/android/diploma/features/favourite/domain/api/FavouriteVacanciesRepository.kt
@@ -1,9 +1,9 @@
 package ru.practicum.android.diploma.features.favourite.domain.api
 
-import kotlinx.coroutines.flow.Flow
+import ru.practicum.android.diploma.features.favourite.domain.model.FavouriteVacancy
 
 interface FavouriteVacanciesRepository {
 
-    fun getFavourites(): Flow<List<Vacancy>>
+    suspend fun addToFavourites(vacancy: FavouriteVacancy, keySkills: String)
 
 }
