+++ app/src/main/java/ru/practicum/android/diploma/features/favourite/domain/interactor/FavouriteVacanciesInteractorImpl.kt
@@ -1,18 +1,15 @@
 package ru.practicum.android.diploma.features.favourite.domain.interactor
 
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 import ru.practicum.android.diploma.features.favourite.domain.api.FavouriteVacanciesInteractor
 import ru.practicum.android.diploma.features.favourite.domain.api.FavouriteVacanciesRepository
+import ru.practicum.android.diploma.features.favourite.domain.model.FavouriteVacancy
 
 class FavouriteVacanciesInteractorImpl(
     private val repository: FavouriteVacanciesRepository
 ) : FavouriteVacanciesInteractor {
 
-    override fun getFavourites(): Flow<List<Vacancy>> {
-        return repository.getFavourites().map { vacancies ->
-            vacancies.reversed()
-        }
+    override suspend fun addToFavourites(vacancy: FavouriteVacancy, keySkills: String) {
+        repository.addToFavourites(vacancy, keySkills)
     }
 
 }
