+++ app/src/main/java/ru/practicum/android/diploma/features/favourite/domain/model/FavouriteVacancy.kt
@@ -3,7 +3,10 @@ package ru.practicum.android.diploma.features.favourite.domain.model
 data class FavouriteVacancy(
     val vacancyId: String,
     val title: String,
-    val salary: String,
+    val salaryFrom: Int?,
+    val salaryTo: Int?,
+    val currencySymbol: String?,
+    val isGross: Boolean?,
     val employerName: String,
     val employerLogoUrl: String,
     val location: String,
