+++ app/src/main/java/ru/practicum/android/diploma/features/favourite/presentation/ui/FavouriteVacanciesFragment.kt
@@ -5,7 +5,6 @@ import android.view.ViewGroup
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import ru.practicum.android.diploma.databinding.FragmentFavouriteVacanciesBinding
 import ru.practicum.android.diploma.features.common.presentation.ui.BaseFragment
-import ru.practicum.android.diploma.features.favourite.presentation.model.FavouriteVacanciesState
 import ru.practicum.android.diploma.features.favourite.presentation.view_model.FavouriteVacanciesViewModel
 
 class FavouriteVacanciesFragment : BaseFragment<FragmentFavouriteVacanciesBinding>() {
@@ -23,12 +22,6 @@ class FavouriteVacanciesFragment : BaseFragment<FragmentFavouriteVacanciesBindin
     }
 
     override fun observeData() {
-        viewModel.getState().observe(viewLifecycleOwner) {
-            render(it)
-        }
-    }
-
-    private fun render(state: FavouriteVacanciesState) {
         TODO()
     }
